backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pypdf==3.17.4
tiktoken==0.5.2
faiss-cpu==1.9.0.post1
numpy==1.26.3
openai==1.10.0
reportlab==4.0.9
python-multipart==0.0.6